feat: Add snackbar component for displaying success and error messages
The code changes include:
- Importing `MatSnackBar` and `MatSnackBarModule` from `@angular/material/snack-bar`
- Adding `SnackbarComponent` to the `entryComponents` array in the `@Component` decorator
- Adding a new HTML template file for the `SnackbarComponent`
- Adding a new SCSS file for styling the `SnackbarComponent`
- Adding a new TypeScript file for the `SnackbarComponent` with logic for closing the snackbar

The purpose of these changes is to implement a reusable snackbar component that can display success and error messages to the user. The snackbar component is used in the `ActivityCreateComponent` and `UserCreateComponent` to show a message when the form is submitted successfully or when there is an error.